requirements.txt
numpy==1.24.2
opencv-python-headless==4.7.0.72
Pillow==9.5.0
streamlit==1.21.0